Viprinet customer router management domains
Viprinet is a manufacturer of VPN Routers. SERIALNUMBER.router.management (.management is a new public TLD) are used to access our routers. In future, we also wish to support Let's encrypt, which requires this domain to be on the public suffix list. Subdomains of the vipri.net domain are internally used for update servers and the likes, and therefore also should be on the public suffic list.