Update for ddnss.de
Please include this domain to the List. I'm not sure if I placed it right. If not, feel free to move it to the right position.
Why the request? I'm using ddnss.de since a long time. It is very reliable and works well with my router.
I recently discovered let's encrypt but unfortunately the rate limit is always reached as more and more people are using it.
I asked the owner of ddnss.de to add the service to this list. In return he asked me to do it on his behalf since his requests have been denied.
Please let me know the reason, should this request be denied as well.