Add certmgr.org
certmgr.org is supposed to become an aiding domain for server management. Any server operator is supposed to be able to obtain a subdomain *.certmgr.org for his server pointing back to the server under the operator's control. In order to avoid the misuse of content on foreign subdomains, strict cookie separation is required.