remove repo conversion .hgtags file (#280)
The .hgtags file is an artifact of the conversion process that pulled
the psl out of a mercurial repository into git.

AFAICT, there's no plan to revert the PSL to mercurial.

So at this point, it's an idle artifact that is more likely to cause
confusion than enlightenment.  We should drop it.