Adding TwoDNS TLD's to public_suffix_list.dat (#328)
In general we have the same needs like other Dynamic DNS Providers. Like others we also have the problem that the owner, for example, of xyz.dd-dns.de is not able to set a cookie for dd-dns.de, which gets transmitted to all users of the specific dynamic 2nd level domain.

In Addition Firefox pulls in a lot of stuff to enhance the user experience for visitors of those sub domains - by using the PSL. So we would like to keep our request up and hope we soon could be added to the list.