Remove 1password domains (#632)
We had a few issues related to PSL. For now it would be easier if we do not have 1password domains listed there.

One of the issues was related to DMARC where emails were not delivered. I believe that FastMail got them resolved but I am worried about other providers running into similar issues. Here is the quote from the email:

> The DKIM alignment check it failing because 1password.com appears on the public suffix list.
>
> RFC7489 (DMARC) section 3.3.1 talks about not allowing domains listed in the psl to align in relaxed mode, and the library we are using for DMARC validation (Mail::DMARC) detects this domain is on the list and returns a fail result for alignment.

We are also not sure if PSL has any effect on Google search results and domain rankings because subdomains are treated differently.