Add YunoHost DynDns domains: nohost.me & noho.st (#615)
YunoHost is a server operating system aiming to make self-hosting accessible to everyone.
We provide dyndns subdomain  `noho.st` and `nohost.me` to our users, allowing them to self-host some services without the need of buying a personnal domain and with the advantage to bypass some ISPs dynamic IP and to configure DNS entries automatically.

It could be very usefull, if browser could correctly manage those subdomains (cookies, highlighting the subdomain, sorting by subdomain and not all the domain ). We want these services to be recognized as a known public suffix and dyndns services too.