Adding own.pm (#544)
We are giving away third level domain for customers of our boxes, and we know that www.ownprovider.com is totally seperated from any other domain like login.ownprovider.com. But for a company which cares about privateness of data and does NOT follow the current "putting all into clouds" and "merging all web apps into a big monster without understanding security side effects or data leakage", this is ok for us. Ownprovider.com is too long for some customers, so we provide also own.pm for customers. Each customer can request a subdomain like cust1.own.pm for free and we don't want to allow change cookies of cust1 for cust2, these 3rd level subdomains should be as seperated as possible. Ownprovider.com should resolve just for safety or short links. It will be redirected to www.ownprovider.com, so cookies are not necessary, and we don't find anything which doesn't allow the top level to resolve.