pcloud.host (#531)
Nucleos Inc is a company behind PortableCloud, a "local cloud solution" (currently something like a platform with an app store so that non-technical users can easily set up whatever software they want on our custom server devices), which is currently expanding its featureset to remote access, supporting SSL certificates, even in LANs and so on. Currently our software is deployed at various schools in Zimbabwe, India and evaluated in countries like Mauritius, rural Mexico and so on.

I can't really give you proper examples due to the nature of our platform (no proper authentication yet and so on), but every device is given its own domain under the pcloud.host suffix (like piotr.pcloud.host, which is currently routed to a tunnelling server, which does not do anything useful yet). The domain is currently mostly used for local access to our devices, upon proper configuration to the network, you can access applications hosted on the device by going to something.[yourdevice].pcloud.host, which has its own HTTPS certificate generated by querying our tunnelling server with a CSR, which is then passed on to Let's Encrypt to generate a cert that contains all DNS names in the SAN field.

Other than that there's the obvious issue of cookies, but we believe that for now it should be enough to only add pcloud.host to the list.