Add telebit.io .app .xyz (#726)
Telebit is similar to services like ngrok, serveo, and localtunnel.me, but we're focused on home and education, and committed to open standards (no vendor lock-in).

Each device gets a semi-random subdomain on one of our domains (such as lucky-duck-42.telebit.io) with unlimited subdomains (i.e. home.lucky-duck-42.telebit.xyz).

We want to remove the rate-limit restrictions of Let's Encrypt as well has have separate cookies for each of the sites.