Proposal to remove bv.nl
I represent the registry for the .nl ccTLD (SIDN) and would like to draw your attention to bv.nl.
To the best of our knowledge that private domain has never properly qualified for the public suffix list, even though there might have been good intentions when the initial request was made. We are not the owner of the private domain bv.nl. We did try to contact the current owner and discuss the matter, but we received no response. Since we are not sure what the correct procedures are in this case, I'd like to ask for extra attention (and help) from the maintainers of this list.