Add site.builder.nu (#723)
site.builder.nu is used for our site builder product. Each customer is assigned a subdomain on *.site.builder.nu -like [randomstring].site.builder.nu.

Customer sites should not be able to share cookies and the URL should be displayed correctly in browsers, and for that reason site.builder.nu should be added to the list.