Add on-rancher.cloud and on-rio.io (#779)
We are adding a feature to automatically provide each user cluster with a <service name>.<random user-id>.on-rancher.cloud (and on-rio.io) DNS entry, so each user-id is a mutually untrusting third-party that should not be able to set cookies readable by each other. We also plan to provide Let's Encrypt wildcard certs for each user-id.