util: replace python newgtlds with newtlds.go
The updated `newtlds.go` util program uses ICANN's v2 JSON registry for
gTLDs instead of the deprecated `newgtlds.csv` data source the legacy
Python utility program used.

The `patchnewgtlds` script is updated to use bash, to fail on pipe
errors, to verify that the `go` binary is in the $PATH, and to use the
new `newtlds.go` program.