bugfix: filter legacy gTLD data to prevent duplicates.
legacyGTLDs are gTLDs that predate ICANN's new gTLD program. These
legacy gTLDs are present in the ICANN_GTLD_JSON_URL data but we do not
want to include them in the new gTLD section of the PSL data because it
will create duplicates with existing entries alongside registry-reserved
second level domains present in the PSL data.