Update public_suffix_list.dat
RLNIC is cooperated with Rusnames (рус/xn--p1acf Registry) in promotion native Russian domains. There are several hundreds of third level domains that are in such program. In order to provide full service for рус/xn--p1acf domains we also provide hosting. Cloudflare is one of the additional services that requires all third level domains in Your list.