Update .so suffix list
In addition to .so, .edu.so, .com.so, .gov.so, .net.so, and .org.so, the .SO ccTLD also supports .sch.so for schools, .me.so for individuals, and .gm.so, .hs.so, .jl.so, .pl.so, and .sw.so for federal member state administrations. I have updated the suffix list to reflect this.