Add gov.scot for Scottish Government
Subdomains of gov.scot are available for registration by other
public sector organisations. For information on availability,
see the Eligibility rules listed here:

  https://www.gov.scot/publications/govscot-domains/

This entry for gov.scot appears in the private domains section of the
list.  This is in contrast to most `gov.*` domains in the list, which
appear in the ICANN section of the list, near or alongside the relevant
country code.  This is because the .scot domain is a (geographic) gTLD
and the gTLD section of the list is auto-generated, so the entry for
`gov.scot` cannot be added alongside the entry for `.scot`. Use of the
private domains section is consistent with, for example,
`service.gov.uk`.