Added senseering domain (#946)
The subdomains will be run by different organisations with the ability to manage the websites and set cookies. To add to this the certificates cant be created by a centralised authority, the participants themselves create them, which could lead to problems with lets encrypt over the entire network instead of for one user.