Add WoltLab Cloud domains (#947)
Within the SaaS solution we will offer a number of domains for customers to choose a subdomain from if they do not want to bring their own domain.

Customers are able to customize the look and feel of their community, including the generated HTML and to a certain extent the server side application code. By doing so they would technically be able to read cookies belonging to the communities of other customers, including HttpOnly cookies. For obvious security reasons we would like to prevent that by submitting the domains to public suffix list.