CI: Add Github Actions workflow for TLD updates PRs.
This commit adds a Github Actions workflow called `tld-update.yml` that
uses `tools/patchnewgltds` to create pull requests from a fork to update
the `public_suffix_list.dat` gTLD section automatically.

It replaces a similar mechanism that used a separate Github repo[0],
a bash script[1], and Travis CI[2] to accomplish the same. Travis has
not been running the job reliably and the Github Action version is more
self-contained, and improves on several other pain points (e.g. it won't
open duplicate PRs).

                      !! Important !!
To operate correctly a secret env var named "BOT_PAT" must be configured
with a Github personal access token for a bot user that has forked the
repository to the same place as configured in `push-to-fork` in
`tld-update.yml`.
                      !! Important !!

[0]: https://github.com/cpu/psl-autopull
[1]: https://github.com/cpu/psl-autopull/blob/master/autopull
[2]: https://travis-ci.com/github/cpu/psl-autopull