Add digitaloceanspaces.com & regional subdomains
Add all regions of DigitalOcean Spaces to the PSL, to deliniate the
names which can be registered by users. This mirrors the inclusion of
Amazon S3.

Please note the contact email address goes to a superset of the domain
registration group, suitable for reaching all critical points of contact
regarding DigitalOcean's PSL entries.

Signed-off-by: Robin H. Johnson <rjohnson@digitalocean.com>