Organize Amazon Domains
Amazon has historically added their domains as per-product domains, inconsistent with other PSL participants.

This groups Amazon domains within one section. This trades the per-product separation, which has benefits, into a better reflection of the number of domains Amazon has, as well as the issues that can arise with the structure of those domains.