Add example and test TLDs
The `example` and `test` TLDs are intended for documentation and testing
respectively (see RFCs 2606 and 6761).  Even though subdomains cannot actually
be registered under these TLDs, if they were real TLDs they would be public
suffixes.  For example, documentation might give `news.example` as an example
of a site, and automated testing might use `news.test` as a test site.
Including these domains on the public suffix matches how they are used.

This PR was prompted by debugging why a browser was not applying cross-site
protection in an automated test which used `site1.test` and `site2.test` as
examples of separate sites.