Remove couk.me and ukco.me from private section (#1519)
We have decided to not renew the domains couk.me and ukco.me, and instead we are focusing on our .uk domains. I wanted to clean these two .me domains up so they aren't left on the list.

These two domains are currently in the expiration grace period, and we would appreciate their removal from the list.

Many thanks.