Create Add lifetime.hosting domains
Purpose of the sites referenced
We provide subdomains (SLDs) to our customers as a cost effective way to maintain a website long term for nominal cost. We use these TLDs as the basis for our client site networks.

Reason for sites to be added
Proper handling of the namespace by browsers. For cookie isolation, SSL management, Direct URL translation, and the inevitable developments in security. To prevent super cookie violation and isolate each SLD domain account from the others in the same namespace. To better isolate malicious content reports and improve our teams response to violated hosting accounts by domain. We are hopeful that this list will help us prevent all of our users from being painted with the same "Malicious code" brush in cases of accounts being violated.
In short, improved security and compartmentalization.