tools: don't render date of contract signature.
ICANN is updating the contract date field in the `gtlds.json` file for
many contracts in the near future. Since including the contract date in
the `.dat` file as comments isn't very helpful to anyone, and leaving it
in during this renewal period would result in a large quantity of
auto-pull requests for no-op changes, we choose to omit the date as
a quick fix.

This commit updates the `newtlds.go` tool to no longer process/render
the `DateOfContractSignature`. Unit tests are updated to reflect this
change.