ci: run Go unit tests as part of `test.yml`.
Previously the Go unit tests for the `newgtlds` tool were run as part of
the Travis workflow for each branch push. This is helpful to catch
breaking changes to the Go tooling _before_ they are merged and used in
the `tld-update.yml` workflow.

This commit adds the required steps to `test.yml` so that these unit
tests are run by GitHub actions similar to the way they were for Travis.