ci: update test workflow triggers to include PRs.
This commit updates the `test.yml` GitHub workflow's triggers to include
`pull_request` in addition to `push`. This will ensure that when a pull
request is opened the CI tasks are run and check statuses added to the
PR.

Without an update to the repository settings to enable branch protection
rules these will not be _blocking_ checks.