Moving priv.at to the ICANN DOMAIN section.
Back when registrations under .at were not possible (early '90ies),
a group of activists created priv.at for domains not associated
with government (gv.at), academia (ac.at), companies (co.at),
or other organisations (or.at). The other domain in this
section (sth.ac.at) is for dormitories attached to the NREN.

See https://en.wikipedia.org/wiki/.at and http://nic.priv.at/geschichte/

Apparently, there is a requirement to add a PSL record to keep
priv.at in the list, as it was added to the PRIVATE DOMAINS
section and not in the ICANN DOMAIN section where it belongs.

Nevertheless, a PSL record will be added to the priv.at zone
once I get the pull request ID of this pull request.

Otmar Lendl (hostmaster@nic.priv.at & lendl@nic.at)