Add closing chevron to contact email address. (#2007)
Conventionally, contact info in the PSL is in RFC 5322 email address
format, aka "Name of Person <email@example.com>". Many entries omit
the chevrons around the email, so parsers need to be robust to that.
But Tor is the only one that is halfway between the two, so it makes
sense to fix it.