Merge WebPros domains in the same section
cPanel, Plesk are part of the same WebPros family
merge all domains together as a single organization entry.

Also note that this change is dropping on purpose the wildcard for cprapid
as this is not needed and we want to be able to store cookies on
x.cprapid.com sub domain.