tools/internal.parser: rework metadata parsing to capture more machine-editable data
Notably, extract all available URLs and emails from lines that can be completely consumed
in an understood format, and collect any incomplete or not understood lines as freeform
"other" text.

Additionally track whether the format seen during extraction suggests that the metadata
could be written back out in a canonical format without destroying any information, so
that a future change can decide whether to try and canonicalize a header and make machine
edits, or whether the metadata is "fragile" and requires human intervention first.