aero: remove extra word between TLD name and URL (#2029)
* aero: remove extra word between TLD name and URL

This is the only TLD that isn't in the form "<tld> : <url>". It's
preferable to make it match the others rather than force metadata
parsers to work around it.

* aero: update domain policy link, and update 2LD public suffixes

The 2LD list comes from clause 11 of the domain management policy.
It is divided into active 2LDs and reserved but inactive. Any active
2LD must provide a registration policy that will be inlined in the
TLD's policies, which is how we can tell the two sets apart.