Add `sslip.io` to private list
sslip.io is a DNS service that, when queried with a hostname with an
embedded IP address, returns that IP address, e.g. 127-0-0-1.sslip.io
resolves to 127.0.0.1.

This fits the category of a domain that "issue subdomains to
mutually-untrusting parties".

Side note: creating the DNS and _PSL TXT style record requires a code
change, so I may try that next time I do a deploy, but that's only every
few months.