tools/internal/github: correctly handle github's mergeability updates (#2214)
Sometimes, the API returns PR info with a null "mergeable" bool, which means
"I need a few seconds to update the PR and check for mergeability, please
try again in a few seconds". Previously this was interpreted by psltool as
"this PR has a merge conflict and cannot be checked".

Instead, do as the Github API docs say, and just keep polling until the PR
information is up to date.