internal/parser: add PublicSuffix and RegisteredDomain methods to List (#2228)
These calculate the public suffix or registered/registerable domain of some
input domain, according to the strict PSL algorithm. Aside from being useful
as a reference implementation of the algorithm, it comes in handy in validations
that need to do stuff like figure out what registry to query for domain info.