Add blogexp.com to the list
Hello,

I'd like to re-submit the request for the addition of blogexp.com to the Public Suffix List.

My previous request(publicsuffix#2564) was closed due to the lack of a valid TLS certificate on the domain. This issue has now been resolved. A wildcard SSL certificate was installed on 20.08.2024, and the site now fully supports HTTPS.

blogexp.com is a platform that allows users to create and manage their own websites with individual subdomains. Each user's site, for example news.blogexp.com or serif.blogexp.com, is administratively under their own control and responsibility.

The primary reason for this request is to ensure proper cookie and security isolation. Since each subdomain is independently managed by its user, treating blogexp.com as a public suffix will prevent one user's actions from affecting another's. This is a critical security measure that helps browsers and applications correctly handle data and sessions for each unique subdomain.

Thank you for your time and consideration.